% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4.R
\name{as.character.md_lines}
\alias{as.character.md_lines}
\title{Convert \linkS4class{md_lines} to \link[base]{character}}
\usage{
\method{as.character}{md_lines}(x, ...)
}
\arguments{
\item{x}{an \linkS4class{md_lines} object}

\item{...}{additional parameters, currently of no use}
}
\value{
The \code{S3} method \code{\link[=as.character.md_lines]{as.character.md_lines()}} returns
a \link[base]{character} \link[base]{vector}.
}
\description{
Convert an \linkS4class{md_lines} object to
a \link[base]{character} \link[base]{vector}.
}
\keyword{internal}
