% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{as_flextable.dist}
\alias{as_flextable.dist}
\title{Convert \link[stats]{dist} into \link[flextable]{flextable}}
\usage{
\method{as_flextable}{dist}(x, ...)
}
\arguments{
\item{x}{\link[stats]{dist} object}

\item{...}{additional parameters of function \code{\link[=as_flextable.matrix]{as_flextable.matrix()}}}
}
\value{
The \code{S3} method \code{\link[=as_flextable.dist]{as_flextable.dist()}} returns a \link[flextable]{flextable}.
}
\description{
..
}
\note{
The \code{S3} method \code{\link[=as_flextable.dist]{as_flextable.dist()}} is inspired by
the function \link[stats]{print.dist}.
}
\keyword{internal}
