% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_duplicated.R
\name{as_flextable.where_duplicated}
\alias{as_flextable.where_duplicated}
\title{\link{as_flextable.where_duplicated}}
\usage{
\method{as_flextable}{where_duplicated}(x, ...)
}
\arguments{
\item{x}{a \linkS4class{where_duplicated} object}

\item{...}{additional parameters, currently of no use}
}
\description{
\link{as_flextable.where_duplicated}
}
