% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibs.R
\name{bib}
\alias{bib}
\alias{.fisher22}
\alias{.pearson1900}
\alias{.holm79}
\alias{.hochberg88}
\alias{.hommel88}
\alias{.benjamini_hochberg95}
\alias{.benjamini_yekutieli01}
\alias{.lawley_maxwell71}
\alias{.tukey49}
\title{\link[utils]{bibentry} of Statistical Methods}
\usage{
.fisher22(key = "Fisher22", ...)

.pearson1900(key = "Pearson1900", ...)

.holm79(key = "Holm79", ...)

.hochberg88(key = "Hochberg88", ...)

.hommel88(key = "Hommel88", ...)

.benjamini_hochberg95(key = "BenjaminiHochberg95", ...)

.benjamini_yekutieli01(key = "BenjaminiYekutieli01", ...)

.lawley_maxwell71(key = "LawleyMaxwell71", ...)

.tukey49(key = "Tukey49", ...)
}
\arguments{
\item{key, ...}{additional parameters of function \link[utils]{bibentry}}
}
\description{
\link[utils]{bibentry} of statistical methods in plain R packages, especially \strong{stats}.
}
\keyword{internal}
