% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_.R
\name{bib_}
\alias{bib_}
\alias{bib_.p_adjust}
\title{\link[utils]{bibentry}-es for R Objects}
\usage{
bib_(x)

\method{bib_}{p_adjust}(x)
}
\arguments{
\item{x}{an R object}
}
\description{
\link[utils]{bibentry}-es for R Objects
}
