% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{clean_char}
\alias{clean_char}
\alias{clean_char.character}
\alias{clean_char.data.frame}
\title{Clean \link[base]{character} Elements}
\usage{
clean_char(x)

\method{clean_char}{character}(x)

\method{clean_char}{data.frame}(x)
}
\arguments{
\item{x}{see \strong{Usage}}
}
\description{
..
}
\keyword{internal}
