% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{data.frame}
\alias{data.frame}
\title{\link[base]{data.frame}}
\description{
\link[base]{data.frame}
}
\examples{
aml2 = survival::aml |>
 within.data.frame(expr = {
  edp = survival::Surv(time, status)
  time = status = NULL
 })
tryCatch(expr = as_flextable(aml2), error = identity)
tryCatch(expr = flextable(aml2), error = identity)
list(
 Formaldehyde = Formaldehyde,
 aml2 = aml2
) |> render2html()

}
