% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factanal.R
\name{factanal}
\alias{factanal}
\title{\link[stats]{factanal} Objects}
\description{
\link[stats]{factanal} Objects
}
\examples{
list(
 factanal = factanal(USJudgeRatings[-1L], factors = 3L)
) |> render2html()

}
