% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{fastmd-package}
\alias{fastmd}
\alias{fastmd-package}
\title{fastmd: Fast Markdown Lines for Common R Objects}
\description{
Fast Markdown lines for common R objects.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
\keyword{internal}
