% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R
\name{htest}
\alias{htest}
\title{\code{htest} Object}
\description{
\code{htest} Object
}
\examples{
list(
 '$t$-test' = t.test(mpg ~ am, data = mtcars)
) |> render2html()

}
\seealso{
\link[flextable]{as_flextable.htest}
}
