% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_}
\alias{inspect_}
\title{Data Inspection}
\usage{
inspect_(x, ...)
}
\arguments{
\item{x}{\link[base]{data.frame}}

\item{...}{additional parameters, currently not in use}
}
\value{
The function \code{\link[=inspect_]{inspect_()}} returns (invisibly) a \link[base]{data.frame}.
}
\description{
..
}
\details{
\itemize{
\item {convert all eligible \link[base]{integer}, \link[base]{numeric},
\link[base]{factor} and \link[base]{character} columns to \link[base]{logical}.}
}
}
\note{
Be aware of potential name clash, e.g., \code{lavaan::inspect}.
}
