% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit.R
\name{matchit}
\alias{matchit}
\title{\link[MatchIt]{matchit}}
\description{
\link[MatchIt]{matchit}
}
\examples{
library(MatchIt)
m = matchit(treat ~ age+educ+race+nodegree+married+re74+re75, data = lalonde)
list(
 '`matchit1`' = m,
 '`matchit2`' = m |> summary(addlvariables = 're78')
) |> render2html()

}
