% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{md_.anova}
\alias{md_.anova}
\title{R Markdown Lines for \link[stats]{anova}}
\usage{
\method{md_}{anova}(x, xnm, ...)
}
\arguments{
\item{x}{an \link[stats]{anova} object}

\item{xnm, ...}{..}
}
\description{
R Markdown Lines for \link[stats]{anova}
}
\keyword{internal}
