% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.stats.R
\name{md_.factanal}
\alias{md_.factanal}
\title{Fast Markdown Lines for \link[stats]{factanal} Objects}
\usage{
\method{md_}{factanal}(x, ...)
}
\arguments{
\item{x}{a \link[stats]{factanal} object}

\item{...}{..}
}
\description{
Fast Markdown Lines for \link[stats]{factanal} Objects
}
\examples{
list(factanal = factanal(USJudgeRatings[-1L], factors = 3L)) |> render2html()
}
\keyword{internal}
