% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.stats.R
\name{md_.htest}
\alias{md_.htest}
\title{Fast Markdown Lines for \code{htest} Object}
\usage{
\method{md_}{htest}(x, xnm, ...)
}
\arguments{
\item{x}{an \code{htest} object}

\item{xnm}{..}

\item{...}{..}
}
\value{
The \code{S3} method \code{\link[=md_.htest]{md_.htest()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Fast Markdown Lines for \code{htest} Object
}
\examples{
list(
 '$t$-test' = t.test(mpg ~ am, data = mtcars)
) |> render2html(file = 'htest')
}
\keyword{internal}
