% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.R
\name{md_}
\alias{md_}
\alias{md_.default}
\alias{md_.list}
\title{R Markdown Lines}
\usage{
md_(x, ...)

\method{md_}{default}(
  x,
  xnm,
  ...,
  fig.height = attr(x, which = "fig-height", exact = TRUE),
  fig.width = attr(x, which = "fig-width", exact = TRUE)
)

\method{md_}{list}(x, xnm, nm_level, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently not in use}

\item{xnm}{..}

\item{fig.height, fig.width}{(optional) \link[base]{double} scalar}

\item{nm_level}{(optional) \link[base]{character} scalar,
mark down level of the \link[base]{list} \link[base]{names},
e.g., \code{'#'}, \code{'##'}, etc.}
}
\value{
The \code{S3} generic function \code{\link[=md_]{md_()}} returns
an \linkS4class{md_lines} object.
}
\description{
To create R markdown lines for various objects.
}
\note{
Read \url{https://plotly.com/r/subplots/}
on how to stack \code{'plotly'} objects
(via function \link[plotly]{subplot}).
}
\examples{
list(
 '`power.htest`' = power.t.test(power = .90, delta = 1)
) |> render2html()

list(
 '`trellis`' = lattice::xyplot(Sepal.Length + Sepal.Width ~ Petal.Length + Petal.Width | Species,
    data = iris, scales = "free", layout = c(2, 2),
    auto.key = list(x = .75, y = .75, corner = c(0.5, 0.5)))
) |> render2html()

list(
 '`ggmatrix`, an `S7_object`' = GGally::ggpairs(swiss, columns = c(1:2, 6))
) |> render2html()

list(
 '`flextable`' = Formaldehyde |> flextable::flextable()
) |> render2html()

list(
 '`magick-image` from package `magick`' = magick::wizard
) |> render2html()

list(
 '`reactable`, an `htmlwidget`' = Formaldehyde |> reactable::reactable()
) |> render2html()

library(leaflet); list(
 '`leaflet`, an `htmlwidget`' = leaflet() |>
  addTiles() |>
  fitBounds(lat1 = 38.85, lat2 = 38.92, lng1 = -77.07, lng2 = -77.0) |>
  addPopups(
    lng = c(-77.0365, -77.0563), lat = c(38.8977, 38.8719), 
    popup = c('white house', 'pentagon')
  )
) |> render2html()

list(
 '`plotly`, an `htmlwidget`' = plotly::plot_ly(z = ~volcano, type = "surface")
) |> render2html()

library(patchwork) # ?patchwork::`patchwork-package`
p1 = ggplot(mtcars) + geom_point(aes(mpg, disp))
p2 = ggplot(mtcars) + geom_boxplot(aes(gear, disp, group = gear))
list(
 '`ggplot`' = p1,
 '`patchwork`' = p1 + p2
) |> render2html()

}
\keyword{internal}
