% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_int.R
\name{md_int}
\alias{md_int}
\alias{md_autoplot_}
\alias{md_grid_draw_}
\alias{md_flextable_}
\title{Fast R Markdown Lines, Internal Use}
\usage{
md_autoplot_(
  x,
  xnm,
  ...,
  summary. = attr(x, which = "summary.", exact = TRUE) \%||\% character(),
  fig.height = attr(x, which = "fig.height", exact = TRUE) \%||\% double(),
  fig.width = attr(x, which = "fig.width", exact = TRUE) \%||\% double()
)

md_grid_draw_(
  x,
  xnm,
  ...,
  summary. = attr(x, which = "summary.", exact = TRUE) \%||\% character(),
  fig.height = attr(x, which = "fig.height", exact = TRUE) \%||\% double(),
  fig.width = attr(x, which = "fig.width", exact = TRUE) \%||\% double()
)

md_flextable_(x, xnm, bibentry. = bib_(x), ...)
}
\arguments{
\item{x}{an R object}

\item{xnm}{\link[base]{character} scalar}

\item{bibentry, summary., fig.height, fig.width, ...}{\link[methods]{slot}s of \linkS4class{md_lines}}
}
\description{
Fast R Markdown Lines, Internal Use
}
\details{
The function \code{\link[=md_autoplot_]{md_autoplot_()}} creates R mark down lines by
the work horse function \link[ggplot2]{autoplot}.

The function \code{\link[=md_flextable_]{md_flextable_()}} creates R mark down lines by
the work horse function \link[flextable]{as_flextable}.
}
\examples{
library(DemographicTable); list(
  '`DemographicTable`' = DemographicTable(CO2, groups = 'Type', include = c('conc', 'uptake'))
) |> render2html()

library(DanielBiostatistics10th); list(
  '`binTab`' = array(c(7L, 3L, 8L, 6L), dim = c(2,2)) |> binTab()
) |> render2html()

}
\keyword{internal}
