% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{not_numeric}
\alias{not_numeric}
\title{Elements which are not \link[base]{numeric}}
\usage{
not_numeric(x)
}
\arguments{
\item{x}{an R object}
}
\value{
The function \link{not_numeric} returns a \link[base]{logical} \link[base]{vector}.
}
\description{
..
}
\details{
The function \link{not_numeric} finds the elements cannot be handled by
\link[base]{as.numeric} (workhorse \link[base]{as.double}).
}
\examples{
not_numeric(c('1.9', '1.1.3', Inf, NA))
}
\keyword{internal}
