% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust_.R
\name{p_adjust_}
\alias{p_adjust_}
\alias{p_adjust_.numeric}
\title{Adjusted \eqn{p}-Values in All Correction Methods}
\usage{
p_adjust_(x)

\method{p_adjust_}{numeric}(x)
}
\arguments{
\item{x}{see \strong{Usage}}
}
\value{
All method dispatches of the \code{S3} generic function \code{\link[=p_adjust_]{p_adjust_()}} return a \link[base]{matrix} of adjusted \eqn{p}-values using all available \link[stats]{p.adjust.methods}.
}
\description{
Adjusted \eqn{p}-values in all correction methods.
}
\examples{
runif(n = 10L, min = .01, max = .2) |>
 p_adjust_() |>
 list('Multile Test Adjustment' = _) |>
 render2html()

}
\keyword{internal}
