% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.htest.R
\name{pairwise_htest}
\alias{pairwise_htest}
\title{\code{pairwise.htest} object}
\description{
The returned object from functions \link[stats]{pairwise.t.test},
\link[stats]{pairwise.wilcox.test} or
\link[stats]{pairwise.prop.test}.
}
\examples{
list(
  '`pairwise.htest`' = airquality |> 
    within.data.frame(expr = {
      Month = factor(Month, labels = month.abb[5:9])
    }) |>
    with(expr = pairwise.t.test(Ozone, Month, pool.sd = FALSE, p.adj = 'none'))
) |> render2html()

}
