% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4.R
\name{print.md_lines}
\alias{print.md_lines}
\title{\code{S3} method \code{\link[=print.md_lines]{print.md_lines()}}}
\usage{
\method{print}{md_lines}(x, ...)
}
\arguments{
\item{x}{an \linkS4class{md_lines} object}

\item{...}{additional parameters, currently of no use}
}
\description{
\code{S3} method \code{\link[=print.md_lines]{print.md_lines()}}
}
\details{
The \code{S3} method \code{\link[=print.md_lines]{print.md_lines()}} is an internal workhorse for
the function \code{\link[=render2html]{render2html()}}.

A use interface, if to be implemented in future, will be in the \code{S4} \code{show} method.
}
\keyword{internal}
