% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render2html.R
\name{render2html}
\alias{render2html}
\title{Create and Render R Markdown file}
\usage{
render2html(
  x,
  path = tempdir(),
  file = stop("must specify `file` explicitly"),
  template = "txz003",
  package = "fastmd",
  rmd.rm = TRUE,
  bib.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{an R object}

\item{path}{\link[base]{character} scalar}

\item{file}{\link[base]{character} scalar}

\item{template, package}{template (from which package) to use; see function \link[rmarkdown]{draft}}

\item{rmd.rm}{\link[base]{logical} scalar, whether to remove the R markdown \code{'.rmd'} file,
default \code{TRUE}}

\item{bib.rm}{\link[base]{logical} scalar, whether to remove the bibliography \code{'.bib'} file,
default \code{TRUE}}

\item{...}{..}
}
\description{
..
}
\keyword{internal}
