% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_call.R
\name{rm_call}
\alias{rm_call}
\alias{rm_call.default}
\title{Remove \code{call} for Printing}
\usage{
rm_call(x)

\method{rm_call}{default}(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Remove \code{call} for Printing
}
\note{
Not all regression models in R suppress the \code{Call} line,
when the \verb{$call} or \verb{@call} information is removed.
}
\examples{
m1 = lm(log(dist) ~ log(speed), data = cars)
m1
rm_call(m1) # 'Call:' still printed

}
\keyword{internal}
