% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4.R
\name{sink2.md_lines}
\alias{sink2.md_lines}
\title{\code{S3} method \code{\link[=sink2.md_lines]{sink2.md_lines()}}}
\usage{
\method{sink2}{md_lines}(x, file, append = TRUE, ...)
}
\arguments{
\item{x}{an \linkS4class{md_lines} object}

\item{file, append, ...}{parameters of the function \link[base]{sink}}
}
\description{
\link[base]{sink} an \linkS4class{md_lines} object to an \code{.rmd} file.
}
\details{
The \code{S3} method \code{\link[=sink2.md_lines]{sink2.md_lines()}},
of the generic function \link[bibentry]{sink2},
is an internal workhorse for
the function \code{\link[=render2html]{render2html()}}.
}
\keyword{internal}
