% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyHSD.R
\name{TukeyHSD}
\alias{TukeyHSD}
\title{\link[stats]{TukeyHSD}}
\description{
\link[stats]{TukeyHSD}
}
\examples{
m = aov(breaks ~ wool + tension, data = warpbreaks)
list(
  'aov' = m,
  '`TukeyHSD`' = m |> TukeyHSD(which = 'tension', ordered = TRUE)
) |> render2html()

}
