% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_duplicated.R
\docType{class}
\name{where_duplicated-class}
\alias{where_duplicated-class}
\alias{show,where_duplicated-method}
\title{\linkS4class{where_duplicated}}
\usage{
\S4method{show}{where_duplicated}(object)
}
\arguments{
\item{object}{a \linkS4class{where_duplicated} object}
}
\description{
\linkS4class{where_duplicated}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{..}

\item{\code{id}}{..}

\item{\code{at}}{\link[base]{character} scalar}
}}

