% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_duplicated.R
\name{where_duplicated}
\alias{where_duplicated}
\alias{where_duplicated.default}
\alias{where_duplicated.data.frame}
\title{\link{where_duplicated}}
\usage{
where_duplicated(x, ...)

\method{where_duplicated}{default}(x, at = "element", ...)

\method{where_duplicated}{data.frame}(x, at = "row", ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently of no use}

\item{at}{\link[base]{character} scalar}
}
\description{
\link{where_duplicated}
}
\examples{
rep(letters[1:3], times = c(2L, 1L, 3L)) |>
 where_duplicated() |>
 list(dup = _) |> render2html()

datasets::Formaldehyde[rep(1:6, times = 1:6), ] |>
 where_duplicated() |>
 list(dup = _) |> render2html()

}
