% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{xtabs}
\alias{xtabs}
\title{\link[stats]{xtabs}}
\description{
\link[stats]{xtabs}
}
\examples{
list(
 '`xtabs`' = xtabs(~ cyl + vs, data = mtcars)
) |> render2html() # citations not shown, probably \pkg{ftExtra} bug

}
