% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.R
\name{md_}
\alias{md_}
\alias{md_.default}
\alias{md_.data.frame}
\alias{md_.xtabs}
\alias{md_.matrix}
\alias{md_.list}
\alias{md_.numeric}
\alias{md_.character}
\alias{md_.noquote}
\title{R Markdown Lines}
\usage{
md_(x, ...)

\method{md_}{default}(
  x,
  xnm,
  ...,
  fig.height = attr(x, which = "fig-height", exact = TRUE),
  fig.width = attr(x, which = "fig-width", exact = TRUE),
  fig.cap = attr(x, which = "fig-cap", exact = TRUE)
)

\method{md_}{data.frame}(x, xnm, ...)

\method{md_}{xtabs}(x, xnm, ...)

\method{md_}{matrix}(x, xnm, ...)

\method{md_}{list}(x, xnm, nm_level, ...)

\method{md_}{numeric}(x, ...)

\method{md_}{character}(x, ...)

\method{md_}{noquote}(x, xnm, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently not in use}

\item{xnm}{..}

\item{fig.height, fig.width}{(optional) \link[base]{double} scalar}

\item{fig.cap}{(optional) \link[base]{character} scalar}

\item{nm_level}{(optional) \link[base]{character} scalar,
mark down level of the \link[base]{list} \link[base]{names},
e.g., \code{'#'}, \code{'##'}, etc.}
}
\value{
The \code{S3} generic function \code{\link[=md_]{md_()}} returns
an \linkS4class{md_lines} object.
}
\description{
To create R markdown lines for various objects.
}
\note{
Read \url{https://plotly.com/r/subplots/}
on how to stack \code{'plotly'} objects
(via function \link[plotly]{subplot}).
}
\examples{
library(lattice); Depth = equal.count(quakes$depth, number=8, overlap=.1)
library(ggplot2)
library(leaflet)
washingtonDC = leaflet() |>
  addTiles() |>
  fitBounds(lat1 = 38.85, lat2 = 38.92, lng1 = -77.07, lng2 = -77.0) |>
  addPopups(
    lng = c(-77.0365, -77.0563), lat = c(38.8977, 38.8719), 
    popup = c('white house', 'pentagon')
  )

library(patchwork) # ?patchwork::`patchwork-package`
p1 = ggplot(mtcars) + geom_point(aes(mpg, disp))
p2 = ggplot(mtcars) + geom_boxplot(aes(gear, disp, group = gear))

list(
 '`power.htest`' = power.t.test(power = .90, delta = 1),
 '`trellis` from package `lattice`' = xyplot(lat ~ long | Depth, data = quakes),
 '`ggplot2::ggplot`' = ggplot() + geom_point(data = mtcars, mapping = aes(wt, mpg)),
 '`GGally::ggmatrix`, an `S7_object`' = GGally::ggpairs(swiss, columns = c(1:2, 6)),
 '`patchwork` from package `patchwork`' = p1 + p2,
 '`flextable::flextable`' = Formaldehyde |> flextable::flextable(),
 '`magick-image` from package `magick`' = magick::wizard,
 '`reactable::reactable`, an `htmlwidget`' = Formaldehyde |> reactable::reactable(),
 '`leaflet::leaflet`, an `htmlwidget`' = washingtonDC,
 '`htmlwidget`' = list(
   plotly::plot_ly(ggplot2::economics, x = ~date, y = ~pop, type = 'scatter', mode = 'markers'),
   plotly::plot_ly(z = ~volcano, type = "surface")
 )
) |> render2html()

aml2 = survival::aml |>
 within.data.frame(expr = {
  edp = survival::Surv(time, status)
  time = status = NULL
 })
tryCatch(expr = as_flextable(aml2), error = identity)
tryCatch(expr = flextable(aml2), error = identity)
list(
 Formaldehyde = Formaldehyde,
 aml2 = aml2
) |> render2html()
list('`xtabs`' = xtabs(~ cyl + vs, data = mtcars)) |> render2html()
list('`matrix`' = VADeaths) |> render2html()

}
\keyword{internal}
