% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_int.R
\name{md_int}
\alias{md_int}
\title{Fast R Markdown Lines, Internal Use}
\usage{
md_int(
  x,
  xnm,
  ...,
  engine = c("print", "autoplot", "grid.draw", "flextable"),
  bibentry. = bib_(x),
  summary. = attr(x, which = "summary.", exact = TRUE),
  fig.height = attr(x, which = "fig.height", exact = TRUE),
  fig.width = attr(x, which = "fig.width", exact = TRUE)
)
}
\arguments{
\item{x}{an R object}

\item{xnm}{\link[base]{character} scalar}

\item{engine}{\link[base]{character} scalar}

\item{bibentry, summary., fig.height, fig.width, ...}{optional, \link[methods]{slot}s of \linkS4class{md_lines}}
}
\description{
Fast R Markdown Lines, Internal Use
}
\details{
To create R mark down lines by
the work horse function
\describe{
\item{\code{engine = 'print'}}{\link[base]{print}}
\item{\code{engine = 'autoplot'}}{\link[ggplot2]{autoplot}}
\item{\code{engine = 'grid.draw'}}{\link[grid]{grid.draw}}
\item{\code{engine = 'flextable'}}{\link[flextable]{as_flextable}}
}
}
\examples{
library(DemographicTable); list(
  '`DemographicTable`' = DemographicTable(CO2, groups = 'Type', include = c('conc', 'uptake'))
) |> render2html()

library(DanielBiostatistics10th); list(
  '`binTab`' = array(c(7L, 3L, 8L, 6L), dim = c(2,2)) |> binTab()
) |> render2html()

}
\keyword{internal}
