% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_int.R
\name{md_int}
\alias{md_int}
\alias{md_autoplot_}
\title{Fast R Markdown Lines, Internal Use}
\usage{
md_autoplot_(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
Fast R Markdown Lines, Internal Use
}
\details{
Function \code{\link[=md_autoplot_]{md_autoplot_()}} creates R mark down lines by
the work hourse function \link[ggplot2]{autoplot}.
}
\keyword{internal}
