% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_int.R
\name{md_int}
\alias{md_int}
\alias{md_autoplot_}
\alias{md_flextable_}
\alias{md_.dist}
\alias{md_.DemographicTable}
\alias{md_.binTab}
\alias{md_.where_duplicated}
\alias{md_.p_adjust}
\title{Fast R Markdown Lines, Internal Use}
\usage{
md_autoplot_(
  x,
  xnm,
  ...,
  fig.height = attr(x, which = "fig-height", exact = TRUE),
  fig.width = attr(x, which = "fig-width", exact = TRUE),
  fig.cap = attr(x, which = "fig-cap", exact = TRUE)
)

md_flextable_(x, xnm, ...)

\method{md_}{dist}(x, xnm, ...)

\method{md_}{DemographicTable}(x, xnm, ...)

\method{md_}{binTab}(x, xnm, ...)

\method{md_}{where_duplicated}(x, xnm, ...)

\method{md_}{p_adjust}(
  x,
  xnm,
  bibentry = c(.holm79(), .hochberg88(), .hommel88(), .benjamini_hochberg95(),
    .benjamini_yekutieli01()),
  ...
)
}
\arguments{
\item{x}{an R object}

\item{xnm}{}

\item{fig.height, fig.width}{(optional) \link[base]{double} scalar}

\item{fig.cap}{(optional) \link[base]{character} scalar}

\item{bibentry, ...}{\link[methods]{slot}s of \linkS4class{md_lines}}
}
\description{
Fast R Markdown Lines, Internal Use
}
\details{
Function \code{\link[=md_autoplot_]{md_autoplot_()}} creates R mark down lines by
the work horse function \link[ggplot2]{autoplot}.

Function \code{\link[=md_flextable_]{md_flextable_()}} creates R mark down lines by
the work horse function \link[flextable]{as_flextable}.
}
\examples{
list(
 USJudgeRatings = dist(USJudgeRatings[1:4,])
) |> render2html()

library(DemographicTable); list(
  '`DemographicTable`' = DemographicTable(CO2, groups = 'Type', include = c('conc', 'uptake'))
) |> render2html()

library(DanielBiostatistics10th); list(
  '`binTab`' = array(c(7L, 3L, 8L, 6L), dim = c(2,2)) |> binTab()
) |> render2html()

}
\keyword{internal}
