% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_int.R
\name{md_int}
\alias{md_int}
\alias{md_autoplot_}
\alias{md_flextable_}
\title{Fast R Markdown Lines, Internal Use}
\usage{
md_autoplot_(
  x,
  xnm,
  ...,
  fig.height = attr(x, which = "fig-height", exact = TRUE),
  fig.width = attr(x, which = "fig-width", exact = TRUE),
  fig.cap = attr(x, which = "fig-cap", exact = TRUE)
)

md_flextable_(x, xnm, ...)
}
\arguments{
\item{x}{an R object}

\item{xnm, ...}{..}

\item{fig.height, fig.width}{(optional) \link[base]{double} scalar}

\item{fig.cap}{(optional) \link[base]{character} scalar}
}
\description{
Fast R Markdown Lines, Internal Use
}
\details{
Function \code{\link[=md_autoplot_]{md_autoplot_()}} creates R mark down lines by
the work horse function \link[ggplot2]{autoplot}.

Function \code{\link[=md_flextable_]{md_flextable_()}} creates R mark down lines by
the work horse function \link[flextable]{as_flextable}.
}
\examples{
list(
 USJudgeRatings = dist(USJudgeRatings[1:4,])
) |> render2html()

}
\keyword{internal}
