% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyHSD.R
\name{TukeyHSD}
\alias{TukeyHSD}
\title{\link[stats]{TukeyHSD}}
\description{
\link[stats]{TukeyHSD}
}
\examples{
m = aov(breaks ~ wool + tension, data = warpbreaks)
list(
 '`aov` ver1' = aov(yield ~  N*P*K + Error(block), npk),
 '`aov` ver2' = list(
  m,
  m |> TukeyHSD(which = 'tension', ordered = TRUE)
 )
) |> render2html()

}
